+++ src/Symfony/Component/PropertyAccess/Tests/PropertyAccessorTest.php
@@ -31,7 +31,6 @@ class PropertyAccessorTest extends \PHPUnit_Framework_TestCase
 
     public function getValidPropertyPaths()
     {
-
         return array(
             array(array('Bernhard', 'Schussek'), '[0]', 'Bernhard'),
             array(array('Bernhard', 'Schussek'), '[1]', 'Schussek'),
@@ -65,7 +64,6 @@ class PropertyAccessorTest extends \PHPUnit_Framework_TestCase
 
     public function getPathsWithMissingProperty()
     {
-
         return array(
             array((object) array('firstName' => 'Bernhard'), 'lastName'),
             array((object) array('property' => (object) array('firstName' => 'Bernhard')), 'property.lastName'),
@@ -86,7 +84,6 @@ class PropertyAccessorTest extends \PHPUnit_Framework_TestCase
 
     public function getPathsWithMissingIndex()
     {
-
         return array(
             array(array('firstName' => 'Bernhard'), '[lastName]'),
             array(array(), '[index][lastName]'),
@@ -129,6 +126,14 @@ class PropertyAccessorTest extends \PHPUnit_Framework_TestCase
     {
         $this->propertyAccessor = new PropertyAccessor(false, true);
         $this->propertyAccessor->getValue($objectOrArray, $path);
+    }
+
+    /**
+     * @expectedException \Symfony\Component\PropertyAccess\Exception\NoSuchIndexException
+     */
+    public function testGetValueThrowsExceptionIfNotArrayAccess()
+    {
+        $this->propertyAccessor->getValue(new \stdClass(), '[index]');
     }
 
     public function testGetValueReadsMagicGet()
@@ -227,6 +232,14 @@ class PropertyAccessorTest extends \PHPUnit_Framework_TestCase
         $this->propertyAccessor->setValue($objectOrArray, $path, 'Updated');
 
         $this->assertSame('Updated', $this->propertyAccessor->getValue($objectOrArray, $path));
+    }
+
+    /**
+     * @expectedException \Symfony\Component\PropertyAccess\Exception\NoSuchIndexException
+     */
+    public function testSetValueThrowsExceptionIfNotArrayAccess()
+    {
+        $this->propertyAccessor->setValue(new \stdClass(), '[index]', 'Updated');
     }
 
     public function testSetValueUpdatesMagicSet()