+++ src/Symfony/Component/PropertyAccess/CHANGELOG.md
@@ -5,6 +5,9 @@ CHANGELOG
 ------
 
  * allowed non alpha numeric characters in second level and deeper object properties names
+ * [BC BREAK] when accessing an index on an object that does not implement
+   ArrayAccess, a NoSuchIndexException is now thrown instead of the
+   semantically wrong NoSuchPropertyException
 
 2.3.0
 ------